SceneManager bug fixes
The initialisation of the first scene in the SceneManager now occurs after the OpenGL ES 'surfaceChanged' call has be made. This allows the width and height of the surface to be accessible from the constructor of a scene. Also added a UI view when trying to run on hardware that doesn't support OpenGL ES 2.0 or higher