Changes to Rendering and some JavaDocs
Made changes to the rendering components and model loader to only support triangle models. This is because the GL_QUADS rendering method is only supported in GLES32 which is not supported by a large amount of android devices. Also improved and completed code checklist for Point2D and Point3D classes